* master:
  Make specex MPI wrapper more robust (#221)
  Quicklook update (#217)
  Dependencies (#220)
  fibermin writing (#216)
  Change the run_steps function so that only the root process reads the graph and checks the file states.  Then broadcast this to all processes.  Also fix a bug where tasks to be computed were marked as in-progress, even if they were already finished.
  Inside this function used by multiprocessing.Pool.map, initialize the return value to None before entering the try/except clause.  This fixes a sporadic hang seen at NERSC.
  also set \$SPECPROD for integration_test
  At debug log level, log the equivalent serial command to run for all tasks that have one.
  Matplotlib and Integration Test Fixes (#213)
  Fix missing import
  Revert "When comparing keys in the QA code, cast to a string representation.  Otherwise the comparison of the exposure ID fails (int vs. string)."
  When comparing keys in the QA code, cast to a string representation.  Otherwise the comparison of the exposure ID fails (int vs. string).
  I/O cleanup (#212)

+++ py/desispec/test/test_image.py
@@ -12,7 +12,7 @@ class TestImage(unittest.TestCase):
         self.pix = np.random.uniform(size=shape)
         self.ivar = np.random.uniform(size=shape)
         self.ivar[0] = 0.0
-        self.mask = np.random.randint(0, 3, size=shape)
+        self.mask = np.random.randint(0, 3, size=shape).astype(np.uint32)
 
     def test_init(self):
         image = Image(self.pix, self.ivar)
@@ -22,14 +22,14 @@ class TestImage(unittest.TestCase):
     def test_mask(self):
         image = Image(self.pix, self.ivar)
         self.assertTrue(np.all(image.mask == (self.ivar==0)*ccdmask.BAD))
-        self.assertEqual(image.mask.dtype, np.uint16)
+        self.assertEqual(image.mask.dtype, np.uint32)
 
         image = Image(self.pix, self.ivar, self.mask)
         self.assertTrue(np.all(image.mask == self.mask))
-        self.assertEqual(image.mask.dtype, np.uint16)
+        self.assertEqual(image.mask.dtype, np.uint32)
 
         image = Image(self.pix, self.ivar, self.mask.astype(np.int16))
-        self.assertEqual(image.mask.dtype, np.uint16)
+        self.assertEqual(image.mask.dtype, np.uint32)
 
     def test_readnoise(self):
         image = Image(self.pix, self.ivar)
